fix: crash on database backup
There was an issue with creating the backup automatically, the reason
for that was that there was a database request locking the file before
doing a backup. This has been resolved with this commit.

Also a warning was added to restore the backup if something went wrong
while deleting the pilot. Additional information was added  to the
disclaimer to help with the process.

fix 61